requirements.txt
pandas
sqlalchemy
pymysql
openpyxl  
python-dotenv  
loguru
